separated dev, demo and production builds
