+++ demo/src/js/index.js
@@ -1,26 +1,11 @@
 import React from 'react';
-import ReactJsonDemo from './components/ReactJsonView/Demo';
+import ReactJsonDemo from './components/Demo';
 
 //index entrypoint component
 export default class extends React.Component {
 
     constructor(props) {
         super(props);
-        this.init(props);
-    }
-
-    state = {}
-
-    defaults = {}
-
-    init = (props) => {
-        for (let i in this.defaults) {
-            if (props[i] !== undefined) {
-                this.state[i] = props[i];
-            } else {
-                this.state[i] = this.defaults[i];
-            }
-        }
     }
 
     render() {
@@ -28,11 +13,6 @@ export default class extends React.Component {
         return (<div class="mac-react" >
             <ReactJsonDemo />
         </div>);
-    }
-
-    componentWillReceiveProps(nextProps) {
-        this.init(nextProps);
-        this.setState(this.state);
     }
 
 }
