+++ docker/debug.sh
@@ -14,7 +14,7 @@ docker run \
     -v $(pwd)/src:/react/src \
     -v $(pwd)/debug:/react/debug \
     -v $(pwd)/entrypoints:/react/entrypoints \
-    -v $(pwd)/webpack.config.js:/react/webpack.config.js \
+    -v $(pwd)/webpack/webpack.config.js:/react/webpack.config.js \
     --rm \
     -e NODE_ENV=$NODE_ENV \
     --entrypoint=/react/entrypoints/debug.sh \
