+++ docker/test.sh
@@ -9,10 +9,9 @@ rm -rf ./coverage/*
 # run the workbench container
 docker run \
     -v $(pwd)/src:/react/src \
-    -v $(pwd)/dev-server:/react/dev-server \
     -v $(pwd)/entrypoints:/react/entrypoints \
     -v $(pwd)/coverage:/react/coverage \
-    -v $(pwd)/webpack.config.js:/react/webpack.config.js \
+    -v $(pwd)/webpack/webpack.config.js:/react/webpack.config.js \
     -v $(pwd)/test:/react/test \
     --rm \
     -e NODE_ENV=$NODE_ENV \
