+++ docker/test_watch.sh
@@ -9,7 +9,7 @@ docker run \
     -v $(pwd)/src:/react/src \
     -v $(pwd)/dev-server:/react/dev-server \
     -v $(pwd)/entrypoints:/react/entrypoints \
-    -v $(pwd)/webpack.config.js:/react/webpack.config.js \
+    -v $(pwd)/webpack/webpack.config.js:/react/webpack.config.js \
     -v $(pwd)/test:/react/test \
     --rm \
     -e NODE_ENV=$NODE_ENV \
