+++ entrypoints/dev-server.sh
@@ -2,6 +2,5 @@
 # runs webpack dev server /react directory
 
 mkdir -p /react/dist || true
-cp /react/src/html/index.html.template /react/dist/index.html
 cd /react
 exec npm run dev:hot
