+++ entrypoints/publish.sh
@@ -3,20 +3,14 @@
 
 export NODE_ENV=${NODE_ENV:-production}
 
-echo Building react app...
+echo Building RJV Distribution...
 
 # remove dist files if they exist,
 # otherwise create the dist directory
-if [ -d /react/dist ]; then
-    echo Removing dist artifacts
-    rm -rf /react/dist/*
-else
-    echo Creating dist directory
-    mkdir /react/dist
+if [ -d /react/dist/main.js ]; then
+    echo Removing existing dist artifacts
+    rm /react/dist/main.js
 fi
-
-# copy over the index.html file for functional test
-cp /react/src/html/test-dist.html.template /react/dist/test.html
 
 # now, build the app
 cd /react
