--- src/html/test-dist.html.template
@@ -1,178 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>react-json-view</title>
-    <meta charset="UTF-8">
-    <!--[if lt IE 9]>
-    <script>
-    (function(){
-      var ef = function(){};
-      window.console = window.console || {log:ef,warn:ef,error:ef,dir:ef};
-    }());
-    </script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.2/html5shiv.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.2/html5shiv-printshiv.min.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/3.4.0/es5-shim.js"></script>
-    <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/3.4.0/es5-sham.js"></script>
-    <![endif]-->
-    <style>
-      .code-comment {
-        color: #aaa;
-        font-size:14px;
-        font-family: monospace;
-        margin-left: 6px;
-      }
-      .react-json-view {
-        padding: 4px 6px;
-        border-radius: 3px;
-      }
-    </style>
-  </head>
-  <body>
-  <div id="app-container">
-    <div class="code-comment">// example 1: JSON object</div>
-    <div class="example" id="example1"></div>
-    <br />
-    <div class="code-comment">// example 2: array input</div>
-    <div class="example" id="example2"></div>
-    <br />
-    <div class="code-comment">// example 3: collapsed by default</div>
-    <div class="example" id="example3"></div>
-    <br />
-    <div class="code-comment">// example 4: base-16 monokai theme</div>
-    <div class="example" id="example4"></div>
-    <br />
-    <div class="code-comment">// example 5: base-16 solarized theme</div>
-    <div class="example" id="example5"></div>
-  </div>
-  <script src="https://unpkg.com/react@15/dist/react.js"></script>
-  <script src="https://unpkg.com/react-dom@15/dist/react-dom.js"></script>
-  <script type="text/javascript" src="./main.js"></script>
-  <style>
-    #app-container>div{
-      margin-bottom: 5px;
-    }
-  </style>
-  <script type="text/javascript">
-      /*-----------EXAMPLE 1-----------*/
-      //some test JSON
-      var test_json = {
-          test: 'click me to expand. I set the `collapseStringsAfterLength` prop to 18 to collapse strings longer than 18 characters.  You can click me again to collapse.',
-          another_sibling: 42,
-          how_will_array_do: [1, 2, 3, 'test'],
-          how_will_floats_do: -2.757,
-          date: new Date(),
-          parent: {
-              sibling1: true,
-              sibling2: false,
-              sibling3: null,
-              'last-sibling': {
-                  grand_child: NaN,
-                  'grand-child-func': (a) => {
-                      let b = a*a;
-                      return b;
-                  }
-              }
-          },
-          string_number: "1234"
-      };
-
-      // reactJsonView is a global variable
-      var jsonView1 = new reactJsonView.default({
-        src: test_json,
-        collapseStringsAfterLength: 18
-      });
-      ReactDOM.render(
-        jsonView1.render(),
-        document.getElementById('example1')
-      );
-
-      /*-----------EXAMPLE 2-----------*/
-      //an example ARRAY
-      var test_array = [
-        'you can also display arrays!', 1, 2, 3, {'pretty_cool': true}
-      ];
-      var jsonView2 = new reactJsonView.default({
-        src: test_array,
-        name: 'array-example'
-      });
-      ReactDOM.render(
-        jsonView2.render(),
-        document.getElementById('example2')
-      );
-
-
-      /*-----------EXAMPLE 3-----------*/
-      var jsonView3 = new reactJsonView.default({
-        name: 'collapsed_by_default_example',
-        src: {
-          example_information: 'this example has the collapsed prop set to true and the indentWidth prop is set to 8',
-          default_collapsed: true,
-          collapsed_array: [
-            'you expanded me',
-            'try collapsing and expanding the root node',
-            'i will still be expanded',
-            {
-              leaf_node: true
-            }
-          ]
-        },
-        collapsed: true,
-        indentWidth: 8,
-        displayDataTypes: false
-      });
-      ReactDOM.render(
-        jsonView3.render(),
-        document.getElementById('example3')
-      );
-
-      /*-----------EXAMPLE 4-----------*/
-      //some test JSON
-      var test_json = {
-          test: 'this is a test string',
-          another_sibling: 42,
-          how_will_array_do: [1, 2, 3, 'test'],
-          how_will_floats_do: -2.757,
-          parent: {
-              sibling1: true,
-              sibling2: false,
-              sibling3: null,
-              'last-sibling': {
-                  grand_child: NaN,
-                  'grand-child-func': (a) => {
-                      let b = a*a;
-                      return b;
-                  }
-              }
-          },
-          string_number: "1234"
-      };
-
-      // reactJsonView is a global variable
-      var jsonView4 = new reactJsonView.default({
-        src: test_json,
-        theme: 'monokai',
-        name: 'monokai-theme'
-      });
-      ReactDOM.render(
-        jsonView4.render(),
-        document.getElementById('example4')
-      );
-
-      /*-----------EXAMPLE 5-----------*/
-      //an example ARRAY
-      var test_array = [
-        'you can also display arrays!', 1, 2, 3, {'pretty_cool': true}
-      ];
-      var jsonView5 = new reactJsonView.default({
-        src: test_array,
-        name: 'solarized-array',
-        theme: 'solarized'
-      });
-      ReactDOM.render(
-        jsonView5.render(),
-        document.getElementById('example5')
-      );
-  </script>
-  </body>
-</html>
