+++ webpack/webpack.config-demo.js
@@ -2,31 +2,18 @@ const path = require('path');
 const webpack = require('webpack');
 
 const PATHS = {
-    src: path.join(__dirname, 'src'),
-    js: path.join(__dirname, 'src/js'),
-    style: path.join(__dirname, 'src/style'),
-    build: path.join(__dirname, 'dist'),
-    demo: path.join(__dirname, 'demo/src')
+    src: '/react/src',
+    js: '/react/src/js',
+    style: '/react/src/style',
+    build: '/react/demo/dist',
+    demo: '/react/demo'
 };
 
 const config = {
-  entry: [PATHS.demo + '/index.js'],
+  entry: [PATHS.demo + '/src/js/entry.js'],
   externals: {
-    'cheerio': 'window',
-    react: {
-      root: 'React',
-      commonjs2: 'react',
-      commonjs: 'react',
-      amd: 'react',
-      umd: 'react',
-    },
-    'react-dom': {
-      root: 'ReactDOM',
-      commonjs2: 'react-dom',
-      commonjs: 'react-dom',
-      amd: 'react-dom',
-      umd: 'react-dom',
-    },
+    'react': 'React',
+    'react-dom': 'ReactDOM',
   },
   output: {
     path: PATHS.demo + '/dist',
@@ -35,7 +22,6 @@ const config = {
     libraryTarget: 'umd'
   },
   plugins: [
-    new webpack.EnvironmentPlugin(['NODE_ENV']),
     new webpack.optimize.UglifyJsPlugin()
   ],
   resolve: {
